Address PR comments and move unit test
